+++ src/formatDateToISO.ts
@@ -3,5 +3,5 @@ import { DateIsoString } from "./types/dateIsoString";
 
 export function formatDateToISO(value: Date | DateIsoString, includeTime: boolean = false): DateIsoString {
   const isoString = convertToDate(value).toISOString();
-  return includeTime ? isoString : isoString.substr(0, isoString.indexOf('T'))
+  return includeTime ? isoString : isoString.substring(0, isoString.indexOf('T'))
 }